requirements.txt
SQLAlchemy==2.0.7
python-dotenv==1.0.0
aiogram==2.25.1
pathlib==1.0.1
requests==2.28.2
loguru==0.6.0
cachetools==5.3.0